port now based on env variable, update to DELETE route

+++ backend/controllers/product.controller.js
@@ -39,6 +39,7 @@ export const updateProduct = async (req, res) => {
     if (!mongoose.Types.ObjectId.isValid(id)) {
         return res.status(404).json({ success: false, message: "Product not found" });
     }
+
     try {
         const updated = await Product.findByIdAndUpdate(id, product, { new: true });
         res.status(200).json({ success: true, data: updated });
@@ -53,11 +54,15 @@ export const deleteProduct = async (req, res) => {
     const {id} = req.params
     console.log("ID to delete: ", id)
 
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ success: false, message: "Product not found" });
+    }
+    
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: "Product deleted" });
     } catch (error){
         console.log("ERROR: ", error.message);
-        res.status(404).json({ success: false, message: "Product not found" });
+        res.status(500).json({ success: false, message: "Server Error" });
     }
 }
